+++ src/chains-config/polkadotControllers.ts
@@ -30,6 +30,7 @@ export const polkadotControllers: ControllerConfig = {
 		'AccountsStakingPayouts',
 		'AccountsValidate',
 		'AccountsVestingInfo',
+		'AhmInfo',
 		'Blocks',
 		'BlocksExtrinsics',
 		'BlocksRawExtrinsics',