+++ src/services/blocks/BlocksService.ts
@@ -37,8 +37,8 @@ import type {
 	WeightV1,
 } from '@polkadot/types/interfaces';
 import type { AnyJson, Codec, Registry } from '@polkadot/types/types';
-import { u8aToHex } from '@polkadot/util';
-import { blake2AsU8a } from '@polkadot/util-crypto';
+import { isObject, u8aToHex } from '@polkadot/util';
+import { blake2AsU8a, decodeAddress, isAddress } from '@polkadot/util-crypto';
 import { calc_partial_fee } from '@substrate/calc';
 import BN from 'bn.js';
 import { BadRequest, InternalServerError } from 'http-errors';
@@ -76,6 +76,7 @@ interface FetchBlockOptions {
 	noFees: boolean;
 	checkDecodedXcm: boolean;
 	paraId?: number;
+	useEvmAddressFormat?: boolean;
 }
 
 interface ExtrinsicSuccessOrFailedOverride {
@@ -119,6 +120,7 @@ export class BlocksService extends AbstractService {
 			noFees,
 			checkDecodedXcm,
 			paraId,
+			useEvmAddressFormat = false,
 		}: FetchBlockOptions,
 	): Promise<IBlock> {
 		const { api } = this;
@@ -238,7 +240,51 @@ export class BlocksService extends AbstractService {
 			finalized,
 			decodedXcmMsgs,
 		};
+
+		const convertToEvm =
+			useEvmAddressFormat &&
+			(this.api.registry.metadata.toJSON().pallets as unknown as Record<string, unknown>[])
+				.map((p) => (p.name as string).toLowerCase())
+				.includes('revive');
+
+		if (convertToEvm) {
+			response.extrinsics = extrinsics.map((ext) => {
+				if (isFrameMethod(ext.method) && ext.method.pallet.toLowerCase() === 'revive') {
+					return {
+						...ext,
+						events: ext.events.map((event) => {
+							return {
+								...event,
+								data: this.convertDataToEvmAddress(event.data.toJSON()),
+							};
+						}) as unknown as ISanitizedEvent[],
+					};
+				}
+				return ext;
+			});
+		}
+
 		return response;
+	}
+
+	private convertDataToEvmAddress(data: AnyJson): AnyJson {
+		// recursive function to convert all AccountId32 to EVM addresses
+		if (data && Array.isArray(data)) {
+			return data.map((item) => this.convertDataToEvmAddress(item));
+			// eslint-disable-next-line @typescript-eslint/no-base-to-string
+		} else if (data && isAddress(data?.toString())) {
+			// eslint-disable-next-line @typescript-eslint/no-base-to-string
+			return u8aToHex(decodeAddress(data.toString()).subarray(0, 20));
+		} else if (isObject(data)) {
+			return Object.entries(data).reduce(
+				(acc, [key, value]) => {
+					acc[key] = this.convertDataToEvmAddress(value);
+					return acc;
+				},
+				{} as Record<string, AnyJson>,
+			);
+		}
+		return data;
 	}
 
 	private async resolveExtFees(