+++ src/services/index.ts
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 export * from './accounts';
+export * from './ahm';
 export * from './blocks';
 export * from './contracts';
 export * from './coretime';